fix: get rid of circular imports in formats
