+++ packages/cli/src/api/formats/index.ts
@@ -1,29 +1,13 @@
 import lingui from "./lingui"
 import minimal from "./minimal"
 import po from "./po"
-import { CatalogType } from "../types"
+import { CatalogFormatter, CatalogFormatOptions } from "./types"
+
+export { CatalogFormatter, CatalogFormatOptions }
 
 const formats: Record<string, CatalogFormatter> = { lingui, minimal, po }
 
 export type CatalogFormat = keyof typeof formats
-
-export interface CatalogFormatOptions {
-  origins?: boolean
-}
-
-export interface CatalogFormatOptionsInternal extends CatalogFormatOptions {
-  locale: string
-}
-
-export interface CatalogFormatter {
-  catalogExtension: string
-  write(
-    filename: string,
-    catalog: CatalogType,
-    options: CatalogFormatOptionsInternal
-  ): void
-  read(filename: string): CatalogType | null
-}
 
 export default function getFormat(name: CatalogFormat): CatalogFormatter {
   const format = formats[name]
