+++ packages/cli/src/api/formats/types.ts
@@ -1,0 +1,19 @@
+import { CatalogType } from "../types"
+
+export interface CatalogFormatOptions {
+  origins?: boolean
+}
+
+export interface CatalogFormatOptionsInternal extends CatalogFormatOptions {
+  locale: string
+}
+
+export interface CatalogFormatter {
+  catalogExtension: string
+  write(
+    filename: string,
+    catalog: CatalogType,
+    options: CatalogFormatOptionsInternal
+  ): void
+  read(filename: string): CatalogType | null
+}
